refactor(auth): move Google auth logic to GoogleAuthService
Removed password generation for Google users
